reorganize modules and graphs into respective directories
