+++ components/dashboard_modules/CurrentOutput.tsx
